// lib/api/models/UserCreate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UserCreate = {
    username: string;
    is_admin?: (boolean | null);
    name?: (string | null);
    surname?: (string | null);
    telegram_id: number;
};